+++ native/storybook/storybook.js
@@ -6,6 +6,7 @@ configure(() => {
   require('../packages/react-native-bpk-component-text-input/stories');
   require('../packages/react-native-bpk-component-text/stories');
   require('../packages/react-native-bpk-component-button/stories');
+  require('../packages/react-native-bpk-component-card/stories');
 }, module);
 /* eslint-enable global-require */